Bill creator component
My first component created with Vue.
This component is compounded by other small components.
It displays a little interface where we can see a little menu and next to it a list with all selected items then these are processed to show their price by unit, discount if it exists, tax for the whole purchase and finally the total mount